ui: Move reaction types out of event_item::content
… into their own submodule.
